index as Anti key pattern
using the index as a key can lead to issues in certain scenarios, especially when the list items are dynamically added, removed, or reordered. This is because React uses the key prop to track the identity of each component within a list, and using the index as a key can produce unexpected results.

When the list changes, React relies on the key prop to determine which components are new, which ones were removed, and which ones were repositioned. Using the index as a key can cause React to mistakenly associate the wrong components when the list changes. This can lead to incorrect rendering, performance problems, and issues with component state and event handling.

To avoid these problems, it's recommended to use stable and unique identifiers as keys instead of relying on the index. Here are a few suggestions for choosing appropriate keys:

Use a unique identifier from your data: If your data items have unique IDs or any other stable and unique identifier, use that as the key. For example, if you have an array of objects with an id property, you can use it as the key: <ChildComponent key={item.id} item={item} />.

Generate unique keys: If your data items don't have unique identifiers, you can generate unique keys dynamically. You can use libraries like uuid or create your own logic to generate unique keys. For example: <ChildComponent key={generateUniqueId()} item={item} />.

Ensure key stability: If the list items are rearranged or updated, ensure that the keys remain stable. This means that the same item should always have the same key, even if its position or content changes.

when to use index as a key:
1. the item in your list do no have a unique id.
2. the list is a static list
3. the list will never be reordered or filtered

2.